{{cookiecutter.project_slug}}/pyproject.toml
[project]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = ""
authors = [{ name = "{{ cookiecutter.author_name }}" }]
requires-python = ">=3.11"

dependencies = [
  "numpy",
  "pandas",
  "geopandas>=1.0.1",
  "matplotlib>=3.10.3",
  "pyyaml>=6.0.2",
  "ruff",
  "typer",
  "tqdm",
  "scikit-learn",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "FLY", "RUF", "B", "ICN"]

[tool.ruff]
line-length = 140
